requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv

# Database
pymongo
motor

# RAG + embeddings
langchain
langchain-community
chromadb
sentence-transformers

# Document processing
pymupdf